requirements.txt
biopython==1.70
numpy==1.13.3
PyQt5==5.9
sip==4.19.3
